removing from "user space" the usage of internal usage method while creating a cli
